Move CLI clarification handler to a separate file
