+++ portia/clarification_handler.py
@@ -1,8 +1,7 @@
 """Clarification Handler.
 
 This module defines the base ClarificationHandler interface that determines how to handle
-clarifications that arise during the run of a plan. It also provides a
-CLIClarificationHandler implementation that handles clarifications via the CLI.
+clarifications that arise during the run of a plan.
 """
 
 from __future__ import annotations
