+++ portia/cli.py
@@ -11,7 +11,6 @@ from __future__ import annotations
 
 import builtins
 import importlib.metadata
-import json
 import sys
 from enum import Enum
 from functools import wraps
@@ -24,7 +23,7 @@ from pydantic import BaseModel, Field
 from pydantic_core import PydanticUndefined
 from typing_extensions import get_origin
 
-from portia.clarification_handler import ClarificationHandler
+from portia.cli_clarification_handler import CLIClarificationHandler
 from portia.config import Config, GenerativeModelsConfig
 from portia.errors import InvalidConfigError
 from portia.execution_context import execution_context
@@ -34,15 +33,6 @@ from portia.tool_registry import DefaultToolRegistry
 
 if TYPE_CHECKING:
     from pydantic.fields import FieldInfo
-
-    from portia.clarification import (
-        ActionClarification,
-        Clarification,
-        CustomClarification,
-        InputClarification,
-        MultipleChoiceClarification,
-        ValueConfirmationClarification,
-    )
 
 DEFAULT_FILE_PATH = ".portia"
 PORTIA_API_KEY = "portia_api_key"
@@ -164,79 +154,6 @@ class CLIExecutionHooks(ExecutionHooks):
     def __init__(self) -> None:
         """Set up execution hooks for the CLI."""
         super().__init__(clarification_handler=CLIClarificationHandler())
-
-
-class CLIClarificationHandler(ClarificationHandler):
-    """Handles clarifications by obtaining user input from the CLI."""
-
-    def handle_action_clarification(
-        self,
-        clarification: ActionClarification,
-        on_resolution: Callable[[Clarification, object], None],  # noqa: ARG002
-        on_error: Callable[[Clarification, object], None],  # noqa: ARG002
-    ) -> None:
-        """Handle an action clarification.
-
-        Does this by showing the user the URL on the CLI and instructing them to click on
-        it to proceed.
-        """
-        logger().info(
-            click.style(
-                f"{clarification.user_guidance} -- Please click on the link below to proceed."
-                f"{clarification.action_url}",
-                fg=87,
-            ),
-        )
-
-    def handle_input_clarification(
-        self,
-        clarification: InputClarification,
-        on_resolution: Callable[[Clarification, object], None],
-        on_error: Callable[[Clarification, object], None],  # noqa: ARG002
-    ) -> None:
-        """Handle a user input clarifications by asking the user for input from the CLI."""
-        user_input = click.prompt(
-            click.style(clarification.user_guidance + "\nPlease enter a value", fg=87),
-        )
-        return on_resolution(clarification, user_input)
-
-    def handle_multiple_choice_clarification(
-        self,
-        clarification: MultipleChoiceClarification,
-        on_resolution: Callable[[Clarification, object], None],
-        on_error: Callable[[Clarification, object], None],  # noqa: ARG002
-    ) -> None:
-        """Handle a multi-choice clarification by asking the user for input from the CLI."""
-        choices = click.Choice(clarification.options)
-        user_input = click.prompt(
-            click.style(clarification.user_guidance + "\nPlease choose a value:\n", fg=87),
-            type=choices,
-        )
-        return on_resolution(clarification, user_input)
-
-    def handle_value_confirmation_clarification(
-        self,
-        clarification: ValueConfirmationClarification,
-        on_resolution: Callable[[Clarification, object], None],
-        on_error: Callable[[Clarification, object], None],
-    ) -> None:
-        """Handle a value confirmation clarification by asking the user to confirm from the CLI."""
-        if click.confirm(text=click.style(clarification.user_guidance, fg=87), default=False):
-            on_resolution(clarification, True)  # noqa: FBT003
-        else:
-            on_error(clarification, "Clarification was rejected by the user")
-
-    def handle_custom_clarification(
-        self,
-        clarification: CustomClarification,
-        on_resolution: Callable[[Clarification, object], None],
-        on_error: Callable[[Clarification, object], None],  # noqa: ARG002
-    ) -> None:
-        """Handle a custom clarification."""
-        click.echo(click.style(clarification.user_guidance, fg=87))
-        click.echo(click.style(f"Additional data: {json.dumps(clarification.data)}", fg=87))
-        user_input = click.prompt(click.style("\nPlease enter a value:\n", fg=87))
-        return on_resolution(clarification, user_input)
 
 
 @click.group(context_settings={"max_content_width": 240})
